requirements.txt
anthropic>=0.84.0
fastapi>=0.135.0
uvicorn>=0.41.0
jinja2>=3.1.0
python-dotenv>=1.2.0
tiktoken>=0.12.0
numpy>=2.0.0
onnxruntime>=1.24.0
tokenizers>=0.21.0
huggingface-hub>=0.30.0
